refactor(MetaPreview): extract useLinksGroups into shared hook
